third_party/blink/web_tests/editing/pasteboard/paste-text-017-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification


Tests: 

Copying and pasting a whole line followed by a blank line could remove the blank line incorrectly, as in the case described in 
<rdar://problem/3951178> REGRESSION (Mail): blank line lost after pasting as quotation.


Expected Results:

Should see a blank line between "two" and "three"


| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <div>
|     "
one
"
|   "
"
|   <div>
|     "
    "
|     <br>
|       class="khtml-block-placeholder"
|     "
"
|   "
"
|   <div>
|     <div>
|       "two"
|   <div>
|     <#selection-caret>
|     <br>
|   "
"
|   <div>
|     "
    three
"
|   "
"
| "
"
